+++ CHANGELOG.md
@@ -11,11 +11,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - PASSBOLT-1917: Upgrade to canjs 3.x
-- PASSBOLT-2883: Fix appjs logout link and remember me cleanup
+- PASSBOLT-2883: Fix logout link and remember me cleanup
 - PASSBOLT-2886: Fix fingerprint tooltips in user group management dialog
 - PASSBOLT-2894: Fix missing div breaking elipsis on long url in password workspace
 - PASSBOLT-2891: Fix group edit users tooltips
-- PASSBOLT-2884: Update header left menu. Remove home and add help.
+- PASSBOLT-2884: Update header left menu. Remove home and add help
+- PASSBOLT-2885: Update user settings menus
 
 # Terminology
 - AN: Anonymous user